Retrieve KEYCODE_ prefix from real android class.
Previously ShadowKeyEvent hardcoded "KEYCODE_" when building the
map of keycode constants. However this value is also stored in a
variable in KeyEvent class. This commit uses reflector to access
the real value. This logic is of course equivalent, but is more
future proof in the unlikely event the keycode prefix changes.

PiperOrigin-RevId: 619980204