Update dependency-on-stubs build.gradle to use AGP
There were several issues with the dependency-on-stubs build.gradle:

1) The original intent of this test is to ensure that the runtime classpath
does not contain the android-all jar, but the android-all was explicitly being
used in the runtime classpath.

2) The android-29 stubs jar was hardcoded in the dependency-on-stubs
build.gradle. It is not guaranteed that android-29 will be installed.

Update the dependency-on-stubs to use AGP, and remove the runtime dependency on
android-all.

Fixes #8962

PiperOrigin-RevId: 622020695