Roll back recent changes to SQLiteDatabaseTest
SQLiteDatabaseTest is designed to only be run in NATIVE SQLite mode.
While some tests may work in LEGACY SQLite mode, many of the tests
cover Android's SQLite extensions, and will never work in LEGACY
SQLite mode.

It is better to disable the ctesque SQLiteDatabaseTest in environments
where NATIVE SQLite mode is not supported.

PiperOrigin-RevId: 624205286