Use Bitmap.copyPixelsFromBuffer to copy image data ImageReader in HW rendering
When HardwareRenderer is used, the underlying Bitmap data is stored as
premultiplied alpha. Previously, when that data is copied to an Android Bitmap
object, Bitmap.setPixels was used, which assumed the int array passed was
un-premultiplied data. Update to use Bitmap.copyPixelsFromBuffer, which does
not make any assumptions about the underlying data. By default Android bitmaps
are stored as premultiplied alpha.

This fixes an issue with alpha blending in HW rendering.

PiperOrigin-RevId: 624739361