Rename `robolectric.screenshot.hwrdr.native` to `robolectric.pixelCopyRenderMode`
The system property `robolectric.screenshot.hwrdr.native` is starting to be
used for some screenshot tests in external Gradle projects. The goal was to
have this be a short-lived internal system property. However, it may need to
persist until HW rendering is more mature (e.g. supported on all RNG SDK
levels) and can support all relevant PixelCopy use cases.

Rename the system property to `robolectric.pixelCopyRenderMode` to make it
clear and unambiguous what the effect of the system property will be. If
`robolectric.pixelCopyRenderMode` is set to `hardware`, HW rendering will be
used for PixelCopy operations. Otherwise, SW rendering will be used for
PixelCopy. Otherwise it has no effect.

PiperOrigin-RevId: 625351004