Add a shadow method for ViewConfiguration.getScaledMaximumDrawingCacheSize
Previously, ViewConfiguration.getScaledMaximumDrawingCacheSize just returned a
constant value, the default max value (1536000 == 480 * 800 * 4).

The result of ViewConfiguration.getScaledMaximumDrawingCacheSize is used in
View.buildDrawingCacheImpl(boolean autoScale), which is used when rendering
View objects.

For tests that configure large screen sizes, the 1536000 may not sufficient for
the cache size. Views that exceed the cache size will not be drawn.

Add a shadow for ViewConfiguration.getScaledMaximumDrawingCacheSize that scales
with the display size.

Also, remove the shadow for ViewConfiguration.getMaximumDrawingCacheSize, the
real implementation can be used instead.

Some tests depend on a minimum cache size of 1536000, such as when they
configure a smaller display and render a large view into it. Add some special
logic that keeps the min cache size at 1536000.

PiperOrigin-RevId: 625381937