Return the correct value in ViewConfiguration.getScaledMinimumScalingSpan
Previously, Robolectric was returning an incorrect value ('0') when
ViewConfiguration.getScaledMinimumScalingSpan was called. This affects
calculations when triggering pinch-and-spread-to-zoom actions.

The real value is defined as '27mm' by default and '32mm' for large
screen devices.

Also add a transient, short-lived system property to allow tests to opt-in
to the incorrect behavior.

PiperOrigin-RevId: 627074303