Use real Android framework code in the ViewConfiguration(Context) ctor
Previously, calling ViewConfiguration.get(Context) did not invoke the
underlying Android framework code. This caused a lot of important
initialization logic to be skipped, and most of the 'scaled' fields in
ViewConfiguration to be uninitialized.

Update ShadowViewConfiguration to shadow the ViewConfiguration(Context)
constructor. This shadow method will call the original real
ViewConfiguration(Context) constructor, and then apply the Robolectric
customizations.

Most of the Robolectric customizations are also incorrect, but these can be
whittled down later.

PiperOrigin-RevId: 627259444