Update ShadowNativeSurface.nativeCreateFromSurfaceTexture for Android V
Previously, ShadowNativeSurface.nativeCreateFromSurfaceTexture invoke the JNI
binding in U and below, but it would return zero in native code, due to
SurfaceTexture not being available on host. This meant that doing the JNI
callback was moot.

In Android V, the native code of Surface.nativeCreateFromSurfaceTexture is not
compiled for host platforms, so '0' has to be returned from the shadow method.

To make these consistent, update
ShadowNativeSurface.nativeCreateFromSurfaceTexture to return zero for all SDK
levels.

PiperOrigin-RevId: 629295331