Set FLAG_HARDWARE_ACCELERATED  in ShadowPackageManger.addActivityIfNotPresent
Previously, only Activities that were present in an AndroidManifest.xml would
contain the FLAG_HARDWARE_ACCELERATED window flag, which gets
populated by PackageParser. This means that Activities not
present in a manifest could not be rendered with a HardwareRenderer, as it
would get ignored due to the missing flag. This meant that the HW rendering
version of PixelCopy would not operate on that Activity.

The FLAG_HARDWARE_ACCELERATED was true by default since
Android ICS (SDK 14), so it should be true by default in Robolectric
as well.

Fixes #8982

PiperOrigin-RevId: 632255888