Stop returning the resources sdk level in RuntimeEnvironment.getQualifiers
Previously, the Build.VERSION.RESOURCES_SDK_INT constant was returned as a
suffix in the call to RuntimeEnvironment.getQualifiers. For example, it could
contain the suffix 'v24'.

There was not much value in returning the SDK level. It was unused, except for
RuntimeEnvironment.setQualifiers checking that the sdk level matched
RuntimeEnvironment.getApiLevel(). It is not possible to set the SDK level using
RuntimeEnvironment.setQualifiers.

RESOURCES_SDK_INT causes problems in the current Android V SDK, as it is higher
than the SDK level. This causes
RuntimeEnvironment.setQualifiers(RuntimeEnvironment.getQualifiers) to throw an
Exception.

Because this API level is unused, it can be removed from the result of
RuntimeEnvironment.getQualifiers.

PiperOrigin-RevId: 632580756