Update some method signatures for Android V
Some graphics static initializers were moved to separate classes in V,
which simplifies the logic to defer static initializers.

Also, add FEATURE_TELEPHONY_SUBSCRIPTION as a default Feature. Without this,
trying to get a VcnManager will return null on V. In general we want telephony
system services to be available in a Robolectric environment.

PiperOrigin-RevId: 632848228