Adds support for new changes to CameraManager/CameraDeviceImpl in V.
The constructor gained a new parameter called CameraDeviceSetup which needed
to be supported. Default to null since it is not required.

Rather than submitting the task and waiting for it, we override the
constructor to set the executor to the one that is given rather than  using
the single thread executor it sets it to. Because Robolectric is quick and
doesn't have a real camera, the background thread is unnecessary complication.

PiperOrigin-RevId: 636813606