Attempt to largely eliminate ShadowBluetoothLeAdvertiser.
ShadowBluetoothLeAdvertiser contains a copy of large portions
of the BluetoothLeAdvertiser code with surprising deviations in
behavior. For example callbacks are made inline rather than
posted to the provided Handler.

This eliminates most of the logic in favor of BluetoothGatt
system service proxy.

Also refactor the IBluetoothManagerDelegates into their own class.

PiperOrigin-RevId: 638342037