Add support for Bitmap.createBitmap(Picture) in SDK 31+ using a Java shim
Previously, calling Bitmap.createBitmap(Picture) in RNG returned null. This is
because the underlying native code,
android_view_ThreadedRenderer_createHardwareBitmapFromRenderNode, is not
supported on host, due to the usage of AImageReader and ANativeWindow NDK
objects.

Use a Java shim that is equivalent but does not use the NDK objects.

This is part of the effort to better support Compose graphics in RNG.

Bitmap.createBitmap(Picture) still does not work in Q and R, but that is due to
a separate set of issues that will be fixed in forthcoming changes.

PiperOrigin-RevId: 639999779