Fix edge case where a ShadowProvider shadow pickers returns null
Previously, if an Android class had a built-in shadow picker, and that shadow
picker returned null, ShadowWrangler would still iterate over the shadow map
and look for a shadow class.

It is more correct not use a shadow if the shadow picker returns null. There is
an implication that the shadow picker is solely responsible for which shadow to
use.

This was causing an issue where RNG native-backed shadows were being used
despite LEGACY graphics mode being configured.

Note that there is also a problem where shadow map entries may appear multiple
times in a generated ShadowProvider. That will be fixed in a forthcoming
change.

PiperOrigin-RevId: 641442109