Add support for delegating proxies with @ClassName.
Previously delegate methods used in ReflectionHelpers#createDelegatingProxy
needed to have their parameter types exactly match the proxy method parameter
types. However that can be problematic for Classes  that do not exist
in the compile-type android all jar. For example, they refer to a Class
from an older SDK that has since been renamed or deleted.

This commit allows the @ClassName annotation to be applied to allow
a custom class name to be applied when looking for a matching method.
And also adds more ReflectionHelper javadoc and unit tests.

PiperOrigin-RevId: 641970912