Make resetState run all ShadowProvider packages even if others fail first.
Previously, a failure in a ShadowProvider's reset would cause any other
packages to not be called. This collects the failing ones and allows the
passing ones to run. In the  future, we may want to completely exit the
test if we find a resetter has an exception thrown as it could cascade
into other tests.

PiperOrigin-RevId: 644075842