Add ShadowDevicePolicyManager#getDelegatedScopes.
ShadowDevicePolicyManager#setDelegatedScopes already added scopes to an internal map,
but ShadowDevicePolicyManager#getDelegatedScopes didn't exist, so it would always return null,
even after delegated scopes were set.
This change adds the corresponding ShadowDevicePolicyManager#getDelegatedScopes
that returns the delegated scopes based on the values in the internal map.

This change also fixes a bug in setDelegatedScopes that would lead to an
UnsupportedOperationException if the function was called multiple times for the same
scope and different packages.

PiperOrigin-RevId: 644590469