Add Shadow implementation for ShadowApplicationPackageManager.getPlatformPermissionsForGroup
This change introduces a limited list of permissions and permission groups
defined in the AOSP platform code. These permissions and permission groups are
used in all PackageManager APIs related to fetching permissions and permission
groups information. You can use the shadow APIs `addPermissionInfo` and
`addPermissionGroupInfo` to add more platform permissions or permission groups,
their name must follow the android convention. These shadow APIs can also be
used to modify the metadata associated with the already defined platform
permissions and permission groups.
This change introduces two changes in behavior. `getGroupOfPlatformPermission`
will return a group only if the input permission is part of this new list of
platform permissions or if its name follows the android convention. And the
test infra will throw an exception if the test app manifest or one of the APIs
to add a package attempt to define a platform permission or permission group.

PiperOrigin-RevId: 649013242