Enable native PixelCopy for Android P
Previously, Android P always used software rendering in PixelCopy.

Add support for HW rendering for PixelCopy in Android P.

Android P currently uses Surface.lockHardwareCanvas to get a hardware
accelerated Canvas. It is not possible to use HardwareRenderer in Android P
(HardwareRenderer was only added in Q). It may be possible to use
ThreadedRenderer using reflection, but that may be overkill. Currently
lockHardwareCanvas does not populate any values for elevation.

PiperOrigin-RevId: 649199544