Bring integration_tests/multidex again with build.gradle
Robolectric has a special module to hook multidex to make it
work with Robolectric, and we need keep these tests for it.

This CL also adds necessary build.gradle for it to make it
work with normal building and testing command.

As Robolectric :shadows:multidex supports android.support.Multidex
and androidx.multidex.MultiDex, this CL add tests for both libraries
to ensure Robolectric's shadowing for these two libraries
work as expected.

Signed-off-by: utzcoz <utzcoz@outlook.com>