Add support for RangingResult.Builder in ShadowRangingResult for V.
In U and below, RangingResult had a large constructor in android.
Robolectric created a builder for this in the shadow. In Android V,
this code now has a builder in the Android code and no longer has a
constructor available. Thus, this change keeps the shadow's builder
and then builds it using the builder in Android. It's not available
for direct usage, so reflection is used to access it.

PiperOrigin-RevId: 651330313