Add tests to `ShadowSharedPreferences`
As reported in #9286, there was an issue in Robolectric < 4.13 when trying to remove a key from `SharedPreferences`.
If one was using `editor.putString("key", null);`, and then check `preferences.contains("key");`,
the result would be `true` instead of `false`.

This has been fixed in Roboletric 4.13.
This PR adds a couple of assertions to prevent regression in the future.