Add a ctesque test for andoid.media.MediaPlayer
ShadowMediaPlayer has state validation logic that was added circa Jelly Bean.
At this point the accuracy of this logic should be scrutinized. Introduce a
ctesque test to allow testing MediaPlayer on both emulators and Robolectric to
ensure consistency. Add an initial test to check that calling
MediaPlayer.getCurrentPosition is allowed in the INITIALIZED state.

Also, add a way to set the error validation mode in ShadowMediaPlayer using a
system property. This allows it to be set more transparently.

PiperOrigin-RevId: 651445737