Add special case handling for NoClassDefError in SandboxTestRunner
On the JVM, if a failure occurs when a class is attempted to be
loaded/resolved, a LinkageError is thrown with detailed information. However,
if the same problematic class is attempted to be loaded again in a subsequent
test, a more truncated and less informative NoClassDefError is thrown.

In Robolectric, this comes up sometimes when the stack trace from the first
test failure is complete and informative, but the stack traces from subsequent
test failures are incomplete.

Add special logic to handle this situation. The first linkage error thrown will
be cached, and subsequent NoClassDefErrors will be compared against the first
LinkageError. If they match, the original LinkageError will be thrown.

PiperOrigin-RevId: 651888069