Add Shadow implementation for ShadowAppOpsManager.setUidMode
In this implementation, the mode set by setMode takes precedence over the mode
set by setUidMode. The real implementation is more complex, certain AppOps can
only be interacted with using the setMode APIs, some only with the setUidMode
APIs, and some with both. For the latter group, it is unclear what the
interaction between the two APIs would be.

PiperOrigin-RevId: 652811043