Add ShadowPermissionControllerManager and shadow revokeRuntimePermissions
The implementation of `revokeRuntimePermissions` is very simple, it just
forwards the call to the PackageManager. The real implementation is part of
mainline and can easily evolve over time. The real implementation upholds the
semantic of different permissions properties by: not revoking permissions that
are considered fixed, revoking more permissions than the caller requested if
they are part of the same group, and more.

PiperOrigin-RevId: 653211678