Update SDK in test resources script and regenerate.
scripts/build-resources.sh is used to build the robolectric test
resources binary. Recent changes broke ShadowActivityTest on gradle
since they made modifications to the test AndroidManifest.xml which
were not present in the binary.

In addition an earlier change that added 'enableOnBackInvokedCallback' prevented
the build-resources.sh from working at all, since that field was added in android
SDK 34, but the script was still using android-28.

This commit modifies the script to use aapt 34 and SDK 34, fail fast on errors,
and regenerates the test resources.

PiperOrigin-RevId: 655747577