Avoid an OOM when PathIterator is used in LEGACY graphics mode
PathIterator was added in Android U and can be used to query
Path objects.

When the PathIterator.nNext native method is called in LEGACY graphics mode, the
default Robolectric return value of zero indicates a VERB_MOVE operation.
This will cause `hasNext` to always return true, which can lead to infinite
loops. Update it to VERB_DONE so hasNext will terminate.

PiperOrigin-RevId: 660934590