Update ShadowVMRuntime.newNonMovableArray to support PathIterator
When it is detected that ShadowVMRuntime.newNonMovableArray is called by
PathIterator, allocate a direct ByteBuffer that can be used in native code.
Note that PathIterator.nNext will still need to be shadowed to update the
backing arrays from native code.

When newNonMovableArray is used from TypedArray, a fake pointer will continue
to be used to avoid the memory overhead of allocating a duplicate array.

Thanks to @brettchabot for the technique.

PiperOrigin-RevId: 662366912