Update the CodeQL workflow
This PR updates the CodeQL workflow to use the latest template provided here:
https://github.com/robolectric/robolectric/new/master?filename=.github%2Fworkflows%2Fcodeql.yml&workflow_template=code-scanning%2Fcodeql

This link can be accessed from https://github.com/robolectric/robolectric/actions/new,
search for "CodeQL", and click "Configure".

The original template had a scheduled trigger to run the workflow.
I removed it, and I restored the `concurrency` rule.

There was also a step to run `./gradlew assemble testClasses`.
Since it is already part of the `tests.yml` workflow, I think it's safe to remove it from here.