Add an public API to get the next vsync time
This is useful for tests that want to have more visibility over the next vsync
event time in order to determine how long to idle the main looper in order to
trigger the next Choreographer frame. This is only used when
Choreographer.isPaused is true (i.e. when the Choreographer is integrated into
PAUSED looper mode).

PiperOrigin-RevId: 666429484