Update ShadowPermissionControllerManagerTest to use a mock OnRevokeRuntimePermissionsCallback
OnRevokeRuntimePermissionsCallback is a hidden API that does not exist in the
stubs jar.

When an anonymous class is used, the Java compiler will create a new class that
extends OnRevokeRuntimePermissionsCallback. This can cause issues when JUnit
scans the class looking for test methods.

This only happens in Gradle environments when the test is run in isolation.

PiperOrigin-RevId: 671830627