Update contribution guide
This PR updates the contribution guide by reusing the content available at https://robolectric.org/contributing/

The only differences between the two documents are:
- Formatting updates.
- New paragraph about Pull Request requirements.

I've also added a new workflow that will comment on the PR once merged,
to remind about the need to redeploy the website.

If this PR is merged, I'll update the website to include this document instead of duplicating it.