Move data listener notification to lower layer as well to cover primitive bytes writes as well.
It would be better to avoid re-implementation for the `write(ByteBuffer)` and
instead only rely on `native` parts, so we use as much as possible of an
original Android `AudioTrack` logic. However doing this would lead multiple
different project tests to fail as they rely on how `ShadowAudioTrack` is
failing where it shouldn't and vice versa. Further, removing re-implementation
of `write(ByteBuffer)` would further require `NioUtils` override via
`Interceptor` and other issues due to some are playing via read-only buffer
where Android has a bug with it. So it would be the best to conduct it if
necessary as a separate step due to affecting area.

PiperOrigin-RevId: 674203166