Add support for NfcAdapter observe mode.
Observe mode is a new NfcAdapter capability being added in android V.
In real android observe mode is stored in a system service, thus the related
APIs don't function as is in Robolectric.

This change adds shadow methods to store observe mode state and test APIs
to control their behavior.

PiperOrigin-RevId: 674340540