Remove generics from ClassName annotation values
ClassName annotation values are designed to work with Class.forName and asm's
Type.getType. These are not allowed to have generics. Remove generic usage from
ClassName values and add a check in the annotation processor for this.

This came up because at Google, there is some custom bytecode instrumentation
that may cause shadow methods to be invoked directly from Android classes, and
it fails when reading ClassName annotation values that are malformed. This is
because the descriptors contained in the ClassName annotation values are used
when generting JVM invocation bytecode.