Extend ShadowStatsManager with addConfig and removeConfig
Add implementations for addConfig and removeConfig to ShadowStatsManager. Also
add method getConfigData to retrieve the statsd configurations stored in the
shadow as a result.

This allows clients of StatsManager to successfully call addConfig and
removeConfig in tests where statsd is not available. It also allows tests to
verify the configurations set into statsd as a result of those calls, which is
otherwise not possible.

PiperOrigin-RevId: 675037915