Remove re-implementation of the `AudioTrack` - `write(ByteBuffer...)`
This method removal in a `ShadowAudioTrack`, enables supporting
of data listener via unified `native_write...` methods, plus
leveraging the most of the real `AudioTrack` implementation.

This required introduction of 2 `NioUtils` methods
`Interceptor`s as they are private to Robolectric.

Further introduce `enableIllegalStateOnPlay` to simulate
an edge when player was in an illegal state while playing.
This allows tests which were depending on `play` to fail
(due to re-implementation of the `write(ByteBuffer...)`
would not adjust the internal state correctly) to behave
the same. As now the internal state of the `AudioTrack`
is correct, `play()` doesn't fail, so we bring that
ability via an explicit opt-in on the `ShadowAudioTrack`.

PiperOrigin-RevId: 678669925