Fix FileDescriptor setFdInt.
Previously the FileDescriptor AndroidInterceptor
used reflection to directly set the 'fd' member field.
In modern versions of the JVM, this logic looks like
it is silently a no-op. This is currently not a problem
with the current BINARY resources implementation, but
is needed for the in-progress NATIVE implementation that
relies on creating a file description in native code.

This commit modified the implementation to use jdk internal
access to set the fd.

This change also requires passing the '--add-opens=java.base/jdk.internal.access=ALL-UNNAMED'
flag to the JVM.

PiperOrigin-RevId: 681950648