Add support for CardEmulationManager observe mode defaults
Observe mode is a new NFC capability being added in android V.
In real android observe mode is stored in a system service, thus the related APIs don't function as is in Robolectric.

This change adds shadow methods to set observe mode defaults and test APIs to check for the defaults.

PiperOrigin-RevId: 684225835