Add a shim method PathIterator.nNextHost
This method can be called instead of PathIterator.nNext. It takes a float array
instead of a memory address. This is more compatible with the JVM, which does
not have non-movable arrays.

Also update the RNR compat library, which fixes an issue with native crashes on
Mac when running multi-sdk tests that include Android V.

PiperOrigin-RevId: 688584888