Adjust to modified BackMotionEvent constructor in in-development SDK.
The latest BackMotionEvent constructor has a new parameter.
This previous logic used the number of constructor parameters
to find a match, which no longer works since there are two variants
with 7 parameters.

This commit changes the logic to use the type of the third parameter
to determine which V+ constructor variant to call.

PiperOrigin-RevId: 688595074