Add basic support for DirectAccessibilityConnection behind a flag
AccessibilityNodeInfo (ANI) is heavily shadowed, and unfortunately many of the
shadow methods are not consistent with real Android. With the current ANI
shadow, it is not possible to use the new
AccessibilityNodeInfo.setQueryFromAppProcessEnabled functionalty introduced in
Android U. This is needed for Compose a11y testing.

Add a system property that defers to the real Android code for ANI. This can
unblock efforts to use Compose a11y testing while not breaking tests that rely
on the existing shadow ANI functinality.

There will be forthcoming commits to clean up Robolectric's ANI shadow and use
more real Android framework code.

PiperOrigin-RevId: 691518699