Deprecate AttributeSetBuilder.
AttributeSetBuilder's implementation is highly coupled to internal resource
implementation and is maintenance prone. Given that Xml.asAttributeSet is an easy,
more future proof replacement for AttributeSetBuilder, deprecate it.

This commit also changes to associated RuntimeEnvironment.compileTimeResTableFile to
be lazy loaded. It is currently only needed when AttributeSetBuilder is used, and
thus lazy loading it will save having to fetch the android-all compile SDK jar.

PiperOrigin-RevId: 692247132