Call AccessibilityNodeInfo.setSource when ShadowAccessibilityNodeInfo.obtain(View) is called
Previously, the AccessibilityNodeInfo.setSource was not being invoked when
ShadowAccessibilityNodeInfo.obtain(View) and
ShadowAccessibilityNodeInfo.obtain(View, int) were called. This means that the
underlying source ids and window ids were not being populated. Update the logic
to call the setSource methods. There will be forthcoming CL to invoke
`obtain(View)` and `obtain(View, int)` directly, but that will require first
updating `obtain()` to be consistent with real Android.

PiperOrigin-RevId: 693204899