Do not require robolectric.useRealAni property to use setQueryFromAppProcessEnabled
Using AccessibilityNodeInfo.setQueryFromAppProcessEnabled requires
that AccessibilityNodeInfo.obtain methods properly populate invoke
the 'setSource' methods. This now happens after the prior
ShadowAccessibilityNodeInfo cleanup changes.

With this change, it is not necessary to specify the robolectric.useRealAni
system property to use AccessibilityNodeInfo.setQueryFromAppProcessEnabled.
ShadowAccessibilityNodeInfo will keep track of whether
setQueryFromAppProcessEnabled was invoked, and use real framework code if
it is invoked.

PiperOrigin-RevId: 693812426