Use applyConfigurationToResources to update resources
This is required to properly update all configurations cached inside the resources
manager, which maintains a weak map of resources keyed on activity keys. When using
activity contexts this is particularly important as otherwise activities can get
created with stale configurations.

This also uses the current configuration that's cached on the activity instance to
compute the configuration change to determine whether to recreate the activity or
not, this aligns with real Android behavior.

Adds the logic flag protected to allow gradual rollout.

PiperOrigin-RevId: 694521783