Remove the `--info` flag from Gradle tasks
This change removes the `--info` flag from Gradle tasks in the CI workflows.
This is done to reduce the verbosity of the output
and make it easier to identify errors.

Note: I left the `--info` flag in the `deploy-snapshot.sh` script.
Let me know if you think I can remove it there too.