Update the Android V SDK to build 12650502
This build contains some fixes for BitmapFactory on Windows. With this build,
the Windows native graphics tests can be re-enabled.

Also, there is logic in hwui that requires the shared library to be named
"libandroid_runtime.dll". So the Android V library naming logic had to be
updated to reflect this.

PiperOrigin-RevId: 696699034