Add realistic support for window insets
Adds support for window insets by implementing behavior of the IWindowSession which
the ViewRootImpl interacts with when adding and modifying its window. The
implementations responds to ViewRootImpl calls and interacts with the IWindow handle
to coordinate the visibility of insets as well as the bounds of the window. This
allows more realistic behavior for more than just window insets, including focus and
window positioning (note that fixed support for window sizing is further flagged to
allow a gradual migration).

Introduces a new SystemUi API for allowing configuring and controlling the window
insets behavior per display. This API is currently not publicly exposed but it is
anticipated once we are happy with the API and behavior this, or another, API will
be made public allowing tests to configure window insets behavior. As window insets
have changed in many Android releases (which contributes to the complexity of
implementing support in Robolectric) its particularly useful to test faithful
behavior across SDKs.

PiperOrigin-RevId: 697697010