Add `ContentResolver#notifyChange(Collection<Uri>, ContentObserver, int)`
This commit adds the missing `notifyChange()` overload in `ContentResolver`,
which takes a collection of `Uri`s.

It also deprecates the `notifyChange(Uri, ContentObserver, boolean)` overload,
since it is deprecated in the Android SDK.

Finally, it changes the `notifyChange(Uri, ContentObserver)` behavior,
by setting the `syncToNetwork` argument to `true`.

Fixes #8633 and #8689.