Lazily initialize member variables in RobolectricTestParameterInjector.DelegateHelperTestRunner
Initializing these variables triggers static initialization of the test class.
This should be deferred as long as possible to ensures it occurs inside of a
sandbox. Before, this wasn't an issue because the HelperTestRunners were being
initialized in the method blocks. However, there will be some forthcoming
changes will cause them to be initialized in the classBlock, before the
sandboxes are initialized.

PiperOrigin-RevId: 718496072