Grouped tests by sandboxes into statement for execution
Setting up Sandbox at class level so all related methods are executed in it.

This changes how tests are run to improve support for `@ClassRule` and reduce
the amount of context sitching that occurred between tests. Previously, all
configured SDK levels would be tested for each individual test before moving on
to the next test. Now, all tests within a specific sandbox are executed
together, and this process is repeated for each sandbox. This update has seen
enhanced performance by reducing the amount of context switching between test
methods noticeably in some of the tests.

For exmaple, given tests T1 and T2 that are configured to run in sandboxes S1
and S2, the previous order was T1S1 -> T1S2 -> T2S1 -> T2S2, where each arrow
is a thread context switch. Now, the order will be (T1S1, T2S1) -> (T1S2,
T2S2). So the number of context switches in this example goes from three to
one. After this change, the number of context switches is equivalent to the
distinct number of sandboxes that are configured in the test.

For #2637

PiperOrigin-RevId: 719451650