Add highly experimental prototype of the Robolectric simulator
This makes it possible to launch a small Swing window to interact with UI
during a Robolectric test, or to use Robolectric to attempt to load an apk file
and a Uber jar. It can help debug why a test may be failing by allowing users
to see what is being displayed during the test. Note this only supports simple
UIs and may have a lot of feature gaps.

It requires some modern modes to be enabled to use (e.g. native graphics).

PiperOrigin-RevId: 722735608