Ignore error logs from FeatureFlagsImpl.
The upcoming in-development version of android
can generate error logs from FeatureFlagsImpl because
that implementation tries to read files that are not
present in the Robolectric environment.
There is no good way to avoid these logs when
running in Robolectric, so this commit modifies
ExpectedLogMessagesRule to ignore those error logs.

PiperOrigin-RevId: 725643982