Remove RobolectricTestRunnerTest.testsWithUnsupportedSdkShouldBeIgnored
This test asserts on behavior that doesn't exist in Robolectric.

It checks the behavior of the test-only StubSdk, which throws an
AssumptionViolatedException when the verifySupportedSdk is called. This test is
unrealistic because the DefaultSdk class that is used will throw an error when
an sdk level is not supported. Remove this test so
RobolectricTestRunner/SandboxTestRunner does not support behavior that doesn't
exist.

PiperOrigin-RevId: 725686874